#65 refactor - repository 인터페이스 추가
