+++ app/src/main/java/com/example/woowagithubrepositoryapp/ui/common/ViewModelFactory.kt
@@ -2,8 +2,8 @@ package com.example.woowagithubrepositoryapp.ui.common
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.example.woowagithubrepositoryapp.data.repository.GithubRepository
-import com.example.woowagithubrepositoryapp.data.repository.TokenRepository
+import com.example.woowagithubrepositoryapp.data.repository.GithubRepositoryImpl
+import com.example.woowagithubrepositoryapp.data.repository.TokenRepositoryImpl
 import com.example.woowagithubrepositoryapp.ui.MainViewModel
 import com.example.woowagithubrepositoryapp.ui.auth.LoginViewModel
 import com.example.woowagithubrepositoryapp.ui.profile.ProfileViewModel
@@ -13,16 +13,16 @@ class ViewModelFactory : ViewModelProvider.Factory {
     override fun <T : ViewModel?> create(modelClass: Class<T>): T {
         return when{
             modelClass.isAssignableFrom(LoginViewModel::class.java) -> {
-                LoginViewModel(TokenRepository.getInstance()) as T
+                LoginViewModel(TokenRepositoryImpl.getInstance()) as T
             }
             modelClass.isAssignableFrom(MainViewModel::class.java) -> {
-                MainViewModel(GithubRepository.getInstance()) as T
+                MainViewModel(GithubRepositoryImpl.getInstance()) as T
             }
             modelClass.isAssignableFrom(ProfileViewModel::class.java) ->{
-                ProfileViewModel(GithubRepository.getInstance()) as T
+                ProfileViewModel(GithubRepositoryImpl.getInstance()) as T
             }
             modelClass.isAssignableFrom(SearchViewModel::class.java) -> {
-                SearchViewModel(GithubRepository.getInstance()) as T
+                SearchViewModel(GithubRepositoryImpl.getInstance()) as T
             }
             else -> throw IllegalAccessException("Unknown ViewModel Class")
         }
